refactor: ts folder
