+++ backend/src/modules/QuestionDetran/ts/enum/QuestionDetranMessagesEnum.ts
@@ -1,0 +1,7 @@
+enum QuestionDetranMessagesEnum {
+    INVALID_TYPE_ERROR = 'Tipo de questão inválido.',
+    INTERNAL_SERVER_ERROR = 'Erro interno no servidor.',
+    QUESTION_NOT_FOUND = 'Questão não encontrada.',
+  }
+  
+export default QuestionDetranMessagesEnum;
