+++ backend/src/modules/QuestionDetran/ts/enum/QuestionOptionsEnum.ts
