+++ backend/src/modules/QuestionDetran/ts/interfaces/QuestionDetranMongooseDocumentInterface.ts
